refactor into dtos packages
