+++ src/main/java/nl/edemtb/mtbclinicsapplication/dtos/mountainbike/MountainbikeInputDto.java
