+++ src/main/java/nl/edemtb/mtbclinicsapplication/mappers/UnregisteredUserMapper.java
@@ -1,0 +1,2 @@
+package nl.edemtb.mtbclinicsapplication.mappers;public class unregisteredUserMapper {
+}
