+++ src/main/java/nl/edemtb/mtbclinicsapplication/services/UnregisteredUserService.java
@@ -1,0 +1,2 @@
+package nl.edemtb.mtbclinicsapplication.services;public class UnregisteredUserService {
+}
